[frontend] refactor button (#13460)
