+++ opencti-platform/opencti-front/src/components/Theme.d.ts
@@ -78,6 +78,41 @@ declare module '@mui/material/styles' {
       dark?: string;
     };
   }
+
+  interface Theme {
+    logo: string | undefined;
+    logo_collapsed: string | undefined;
+    borderRadius: number;
+    button: {
+      sizes: {
+        default: SizeConfig;
+        small: SizeConfig;
+      };
+    };
+  }
+
+  interface ThemeOptions {
+    logo?: string | null;
+    logo_collapsed?: string | null;
+    borderRadius?: number;
+    button?: {
+      sizes?: {
+        default?: SizeConfig;
+        small?: SizeConfig;
+      };
+    };
+  }
+}
+
+export interface SizeConfig {
+  height: string;
+  padding: string;
+  minWidth: string;
+  width: string;
+  fontSize: string;
+  fontWeight: number;
+  lineHeight: string;
+  iconSize: string;
 }
 
 interface ExtendedColor extends PaletteColorOptions {
@@ -128,6 +163,12 @@ interface ExtendedThemeOptions extends ThemeOptions {
   logo_collapsed: string | null;
   palette: ExtendedPaletteOptions;
   borderRadius: number;
+  button: {
+    sizes: {
+      default: SizeConfig;
+      small: SizeConfig;
+    };
+  };
 }
 
 export interface Theme extends MuiTheme {
@@ -135,4 +176,10 @@ export interface Theme extends MuiTheme {
   logo_collapsed: string | undefined;
   borderRadius: number;
   palette: ExtendedPaletteOptions;
+  button: {
+    sizes: {
+      default: SizeConfig;
+      small: SizeConfig;
+    };
+  };
 }
