+++ opencti-platform/opencti-front/src/components/common/button/Button.types.ts
@@ -1,39 +1,6 @@
-import { ButtonProps as MuiButtonProps } from '@mui/material';
-
-export type ButtonVariant = 'primary' | 'secondary' | 'tertiary' | 'extra';
 export type ButtonSize = 'default' | 'small';
 export type ButtonIntent = 'default' | 'destructive' | 'ai';
 export type GradientVariant = 'default' | 'ai' | 'disabled';
-
-export interface BaseButtonProps extends Omit<MuiButtonProps, 'variant' | 'color' | 'size'> {
-  variant?: ButtonVariant;
-  intent?: ButtonIntent;
-  size?: ButtonSize;
-  gradient?: boolean;
-  gradientVariant?: GradientVariant;
-  gradientStartColor?: string;
-  gradientEndColor?: string;
-  gradientAngle?: number;
-  startIcon?: React.ReactNode;
-  endIcon?: React.ReactNode;
-  fullWidth?: boolean;
-  iconOnly?: boolean;
-  component?: React.ElementType;
-  to?: string
-}
-
-type RestrictedIntentButtonProps = BaseButtonProps & {
-  intent: 'destructive' | 'ai';
-  variant?: Exclude<ButtonVariant, 'primary'>;
-};
-
-// Default buttons can use any variant
-type DefaultIntentButtonProps = BaseButtonProps & {
-  intent?: 'default';
-  variant?: ButtonVariant;
-};
-
-export type CustomButtonProps = RestrictedIntentButtonProps | DefaultIntentButtonProps;
 
 export interface ColorDefinition {
   main: string;
@@ -47,4 +14,15 @@ export interface ColorDefinition {
 export interface GradientColor {
   start: string;
   end: string;
+}
+
+export interface SizeConfig {
+  height: string;
+  padding: string;
+  minWidth: string;
+  width: string;
+  fontSize: string;
+  fontWeight: number;
+  lineHeight: string;
+  iconSize: string;
 }
