+++ opencti-platform/opencti-front/src/public/components/LoginForm.tsx
@@ -1,7 +1,6 @@
 import React, { FunctionComponent } from 'react';
 import { Field, Form, Formik } from 'formik';
 import { TextField } from 'formik-mui';
-import Button from '@mui/material/Button';
 import { graphql } from 'react-relay';
 import * as Yup from 'yup';
 import { FormikConfig } from 'formik/dist/types';
@@ -10,6 +9,7 @@ import { useTheme } from '@mui/styles';
 import { Theme } from '@mui/material/styles/createTheme';
 import { useFormatter } from '../../components/i18n';
 import useApiMutation from '../../utils/hooks/useApiMutation';
+import { Button } from '@mui/material';
 
 const loginMutation = graphql`
   mutation LoginFormMutation($input: UserLoginInput!) {
@@ -101,7 +101,7 @@ const LoginForm: FunctionComponent<LoginFormProps> = ({ onClickForgotPassword, e
               style={{ marginTop: theme.spacing(3) }}
             >
               {t_i18n('Sign in')}
-            </Button>
+            </Button>             
           </Form>
         )}
       </Formik>
